Update sources structure
