+++ include/i_messages_processor.h
@@ -1,0 +1,8 @@
+#pragma once
+
+class MessagesProcessor {
+public:
+  virtual ~MessagesProcessor() {}
+
+  virtual void checkProcessedMessages() = 0;
+};
