+++ net/session_manager.cpp
@@ -1,0 +1,13 @@
+#include "net/session_manager.h"
+
+void SessionsManager::addSession(MessagesProcessorPtr processor) {
+  processors.insert(processor);
+}
+void SessionsManager::removeSession(MessagesProcessorPtr processor) {
+  processors.erase(processor);
+}
+void SessionsManager::checkProcessedMessages() {
+  for (auto processor : processors) {
+    processor->checkProcessedMessages();
+  }
+}
